Thread popt through parser heap, load from disk if AST is not present
Summary: This allows us to lazily load ASTs from the DiskHeap, which will be necessary when loading from saved state when we don't end up parsing our entire codebase. This also causes the disk heap to try to load from actual disk if it can't find it in shared memory.

Reviewed By: alexchow

Differential Revision: D4054538

fbshipit-source-id: 579762bdbd8833c77f6477e966b1ce6057274826