Fix returnSPOffset handling for InterpOnes
Summary:
When we spill a frame for a call, we store the `returnSPOffset` in
frame-state's FPI stack. This is the offset of the bytecode SP when the call
returns and is used when inlining. However, if we InterpOne a FPush* op, we
store the IR SP as the offset instead of the BC SP. These don't necessarily
match. Moreover, we fail to account that the InterpOne may pop values off the
stack before spilling the frame. CufIterSpillFrame has a similar issue. Fix this
by passing the BC SP as intended.

This error is somewhat benign, as the returnSPOffset is only used when inlining,
and we won't inline function calls which spill the frame using InterpOne or
CufIterSpillFrame (because we don't know the callee). It can however cause
assertions to trigger if there's a path which uses InterpOne to spill the frame,
and one that doesn't. Where they join, the FPI stack in frame-state won't agree.

Reviewed By: swtaarrs

Differential Revision: D4174556

fbshipit-source-id: f83705d8836635342e442645d248231499545e55