Add an option to sort translations by "density" of perf events
Summary:
With the region JIT, we often end up with very large translations, which may
have a bunch of perf samples because of their size.  This diff adds a new option
allows tc-print to sort the top translations by density of the events, which
hopefully can help us find smaller translations with interesting performance
characteristics.

Reviewed By: paulbiss

Differential Revision: D4180502

fbshipit-source-id: 02c88ed612dcaf538730ab0c47b8a3c0cc12b3fb