Convert remainder of ObjectData scanning to type_scan
Summary:
Scan objects with type_scan::Scanner instead of the old infra.
Also refactored ObjectData::scan() so it only scans the generic
part of ObjectData that's present for all variations of layout.
The kinds that have special layout also have special HeaderKinds,
so now we dispatch on kind, do the special stuff, then use OD::scan
for the generic part.

Reviewed By: jano

Differential Revision: D4023578

fbshipit-source-id: 66cd5bb85a01367ad21a733ec2660774d7d12fd7