v0.2.9 - Binary Expression formatting and test
Summary:
The key insight into binary expression formatting is realizing that there are different categories of binary expressions and we (as humans formatting code) have different conventions for them.

We have special rules for indentation that apply to assignment and comparison operators that don't apply to any other operations so we treat them separately.

For these other operators, we keep each operand on the same nesting level, as long as the precedence of the operators is the same.
We do this by walking the tree of potential binary operations and turning it into a list. Each operand in the list represents where a potential line break can occur, and we can also nest the expression as needed.

The other complication is that for binary expression with different varieties of precedence, we need to know that we are in a sub binary expression, so we can nest after the first operand. This is meant help the reader understand the order of evaluation. For example, in the expression $a + $b * $c, we format it on multiple lines as

  $a +
  $b *
    $c

to indicate that the $b * $c part is evaluated first.

Differential Revision: D4178920

fbshipit-source-id: 1aa6b8dac2f28236d14462a378d9c8b057ef728d