Stash type index in req::make_shared; scan by index.
Summary:
An alternative to conservative scanning of req::shared_ptr.
Because a shared_ptr<T> can be cast to a derived type D later, stash the proper size/index at the time of instantiation,
which right now is limited to make_shared.

Questions:
- Instantiating a req::shared_ptr the default way (req::shared_ptr foo;) is still valid, calls through to shared_ptr().
How to handle this case - should I check for valid values in TYPE_SCAN_CUSTOM? Is it fine as is? Make shared_ptr() private?

Depends on D4135678

Reviewed By: ricklavoie

Differential Revision: D4163689

fbshipit-source-id: 64bdea013c1d0063742e4ecf247b4007ed1f4927