Handle FPushCtorI in func-effects.cpp
Summary:
We weren't handling FPushCtorI when determining if a function call can destroy
the caller's locals. As a result, we were doing the conservative thing and
assuming that it could. It can't, so handle it like the rest of the constructor
calls. Add an assert for the default case so we'll catch any such future changes
immediately.

Reviewed By: markw65, swtaarrs

Differential Revision: D4187781

fbshipit-source-id: a0addc1270311aa861560d19d91fe8ac11fc0a88