Fix bug in EditableSyntaxList iterator, add tree dumping sample
Summary:
I've created a sample that shows how to create a tree dumper that takes in source code and produces nicely formatted parse trees using box-drawing characters.

While writing that I found a bug -- that I had already marked with a TODO, silly me -- in how children of a list are iterated. I fixed it.

Reviewed By: dlreeves

Differential Revision: D4184943

fbshipit-source-id: d343f1133920562e237c9d063790fc7a4ad86259