Inherit from ActRec to clue in type-scanning
Summary:
Enable scanning of ActRec* fields by using inheritance.
Pointers to bases of heap-allocated types are "interesting"
and get scanned by default. We currently require heap-allocated
ActRecs to be the first field in their respective container
types (Resumable, BaseGenerator, etc), but this isn't enough
to make ActRec* be scanned when it shows up in other indexed
types. And, inheritance is a decent way to ensure the layout
we require anyway.

Reviewed By: jano

Differential Revision: D4160419

fbshipit-source-id: 4f0b7a49216d6120fd439450efcd36d5249e10af