Added file_with_modes object, store it in saved state
Summary:
In order for lazy parsing to work, we need the file mode of every file, which isn't saved in our fast object. This adds a fast_with_modes object which essentially just stores a file mode with every set of toplevel names.

This doesn't increase the saved state size significantly(< 1MB before compression), and makes it a lot cheaper than reading and parsing modes on the fly.

(Flaky tests, so overriding unit failures)

Reviewed By: andrewjkennedy

Differential Revision: D4085054

fbshipit-source-id: 4a86541c5f504632df2bd031c7342b4cd2bed9f5