Load parsing from saved state
Summary:
So the entire point of this entire stack of diffs is to give us the infrastructure we need to lazily parse. This diff accomplishes this goal.

From the saved state, we load a file_with_modes object. This object contains the modes and toplevel names of all files in our codebase. We then parse and name only the changed files.

Then, we name_with_fast all of the files from our saved state, and add them to our environment's file_info object.

Things missing from this diff at the moment:
- Loading IDE indices. I'll be adding this in the diff above
- Gating so we don't release this to all users. Will be amended into the top diff once it's done.

I figured it's best to get this one out there anyway, to give some context for the previous diffs.

Reviewed By: dlreeves

Differential Revision: D4086037

fbshipit-source-id: 332e001789b8b86527c7e508c7bb74c20e4ee99f