Made search service great again
Summary:
This diff adds updating the search service back to initialization, since it was removed in the diff before.

That said, we are missing some information when loading from saved state: class kinds and positions aren't loaded, and we have no methods information. We fix the first of these categories by lazily loading class kinds and positions of our search results. Loading all class methods is a much more difficult task, since it inherently depends on the size of the codebase, so I leave it out for now.

Reviewed By: jopamer

Differential Revision: D4118927

fbshipit-source-id: 34882aacdf38f4d3e00c4a1b6f5589230e80cba0