The real fix
When exiting the VM, the r1 register is read in order to reload the r2
(toc register). Other than that, only the r31 is used as a frame pointer
in order to not only read from stack but only to pop the frame.

The TOC register had the same data before toc/relocation PR (7389) so
this bug was not affecting when fix backchain PR (7327) was integrated.