Add logging for stale file heap
Summary:
If one does several large(>5000 files) rebases in quick succession with lazy parsing on, theres a chance that the files on disk will change while we typecheck before Watchman can tell us about it: this can cause issues where our FileInfo object is now stale, and our file_heap no longer represents the best we know about the disk.

This diff catches this error, which, due to Watchman Subscriptions should be extremely rare, and gracefully restarts the server.

Because we want to be headed toward a world where we reinitialize on large rebases anyway, this is a reasonable thing to do. Incremental mode is most likely slower than a full server init for giant rebases, and therefore restarting the server may actually save the user time.

Reviewed By: alexchow

Differential Revision: D4201531

fbshipit-source-id: a70a2d0483115ca70c8ef3df41a8f79610bb8d5d