Fix type-scanner generation for Resumable when it has ActRec as a base
Summary:
The members of an union can be in an indeterminate order in the debug
information, so any logic which relies on their ordering can have problems.

If Resumable has ActRec as its base, we'll attempt to generate inline logic to
scan the ActRec base (we don't generate a scanner for ActRec otherwise). ActRec
has two unnamed unions as members, with annotations to conservative scan
both. However, the annotations refer to a particular member of each unnamed
union (its the only way since the unnamed unions can't be referenced otherwise),
and the second union has all "interesting" types as its member. So, if we
process the members without the annotation first, we'll generate layout for
those. When we finally hit the member with the annotation, we'll generate a
conservative scan for the union. As a result, the generated scanner will enqueue
the same offset twice, and then conservative scan it for good measure.

Fix this by only attempting to generate the layout for an union until we've
iterated over all the members of that union and verified that there's no special
annotations for any of them. In addition there were some bugs with calculating
some offsets, so fix those too.

Differential Revision: D4225329

fbshipit-source-id: f88008c75cb894a62fefb7bffd60b54d4214deca