Fix use-after-free
Summary:
asan tests revealed a use-after-free due to D4145828. I had tried to
match php5 behavior in that diff, but it turns out there's no good way
to do that, because there's no way to protect "base" from a malicious
error handler. Matching php7 sidesteps the issue, and the difference
is very hard to observe.

Reviewed By: swtaarrs

Differential Revision: D4227027

fbshipit-source-id: 7bbe683614d95d00308b576884d6b0a9b4723693