Fix TOC register restore due to new backchain fix on PPC64
Summary:
The TOC register restoration was being handled in 2 different ways by 2
pull requests (7389 and 7327) but together there were 7 tests from the
quick test suite that started failing.

When exiting the VM, the r1 register is read in order to reload the r2
(toc register). Other than that, only the r31 is used as a frame pointer
in order to not only read from stack but only to pop the frame.

The TOC register had the same data before toc/relocation PR (7389) so
this bug was not affecting when fix backchain PR (7327) was integrated.
Closes https://github.com/facebook/hhvm/pull/7498

Differential Revision: D4202889

Pulled By: mxw

fbshipit-source-id: a88af70ec05dc8f0d589373b9853e47e01f81e3b