Support contravariant subtyping for function type arguments
Summary:
Hack function types such as `(function(arraykey):int)` currently behave covariantly
in return types (so this type is a subtype of `(function(arraykey):arraykey)`) but invariantly
in argument types. It's sound -- and more flexible -- to permit *contravariance* in
argument types, so the type above is a subtype of `(function(int):arraykey)`.

Reviewed By: jopamer

Differential Revision: D3834381

fbshipit-source-id: 3c5777f05dc0fa8f6b8a39c553e61fdb4db2e053