Use size_t for jit size
Summary:
pcre_fullinfo takes a void*, and uses it to return many different
types. jitted code size is returned as a size_t, not an unsigned int.

We seem to be getting away with this currently, but in lto builds its
overwriting the bottom bits of spattern, causing a segfault on exit
from pcre_get_compiled_regex_cache.

Reviewed By: ricklavoie

Differential Revision: D4237827

fbshipit-source-id: 9495ddb8f89c52b30152c327ae1ed8c19c3abfbc