Filter out the method when identifying its parameters.
Summary:
Currently, identifying the "bar" in "foo(bar)" also gives a result for "foo".

This change filters out the method when the item to be identified is a parameter. It does this through the not-so-awesome approach of identifying whether the cursor is positioned before the first "(" in the string that the method span covers in the source code.

Reviewed By: dabek

Differential Revision: D4224287

fbshipit-source-id: 46bb88cfd4bbd6f8ba1b8c1d14dab2f351a2af38