Add EvalEnableContextInErrorHandler option
Summary:
Add an option called EvalEnableContextInErrorHandler to control whether
we pass a "context" array to a registered error handler or not. The context
array contains a copy of all the locals in the function where the error was
raised. If the option is not set, than an empty array is always passed. Passing
such a context can inhibit JIT optimizations because it prevents us from eliding
stores to locals which would not otherwise have to be read. The option is off by
default in RepoAuthoritativeMode.

Reviewed By: markw65

Differential Revision: D4241094

fbshipit-source-id: a780d6dde352b0e7319bf1aeb3cba0e7ef0e9f82