Use AttrIsFoldable in HHBBC's builtin constant folding
Summary:
We have AttrIsFoldable now, so rather than use a hard-coded list of builtins
safe to constant fold, use that instead. Mark a few additional builtins with
AttrIsFoldable which HHBBC considered safe to fold. While we're at it, do a
little bit of cleanup. Remove the dedicated readsCallerFrame and
writesCallerFrame variables from php::Func and just check the attributes
directly. Also treat sizeof as a synonym for count in irgen-builtin.cpp.

Reviewed By: markw65

Differential Revision: D4214886

fbshipit-source-id: 53cd3644af477414f25c7260761cebd8bf924d2f