Rename some stack offsets
Summary:
Our naming system for stack offsets is generally terrible.  Rather than
describing what the offset represents, the names we use instead describe
what coordinate system the offset is in.  But that's what we have the
various stack offset types for---and in most contexts, what the offset
is representing (i.e., the offset of the IRSP relative to the FP, the
offset of the BCSP relative to the IRSP, etc.) is not self-evident.

This diff changes some of the names, but not all of them because I
started to fall asleep.

Reviewed By: markw65

Differential Revision: D4186873

fbshipit-source-id: 198beefea562675d9aa680210e53a030dfd8e7fe