bser test harness
Summary:
Preliminary stuff for an ocaml implementation of bser,
the binary format used by watchman.

Some boilerplate to parse command-line args and convert
between formats (bser and json) for testing.

Meaning:

* `--to_bser` take json input and output the corresponding
  bser

* `--to_json` take bser binary data as input and output
  the corresponding json

* `--roundtrip` take bser binary data as input, parse it,
  convert back to bser and output (?)

The `--roundtrip` option is there to test the 'template'
facility in BSER (a space efficient way to represent
a table of objects with the same fields). There is no
json way of representing 'I want this to be a templated
table in the BSER output,' so the idea is to just record
parser events.

Reviewed By: jopamer

Differential Revision: D3809350

fbshipit-source-id: 012d8d11729d36e4f2fd90c21b39462778a953eb