parse strings and objects
Summary:
See format at https://facebook.github.io/watchman/docs/bser.html

Similar to arrays, but now requiring strings for field names. Objects
are encoded

  [tag][number of fields][fields]

Where the tag is `\x01` and each field is of the form:

  [name][content]

And each name is a string with form

  [tag][length][content]

where the tag is `\x02`, the length is a BSER integer, and the
content is just a byte array.

Reviewed By: mbouaziz

Differential Revision: D3809383

fbshipit-source-id: 09d91465036e22ad5d713a2e846795866331e1f3