convert exceptions to carry stream position
Summary:
Decided this would be more useful than repeating the `__LOC__`
(which ideally should be showing up in a stack trace already).

Carry the byte offset into the input instead, which is useful
for debugging. Downside: forced to use some sentinel in places
where I don't have access to the stream; using `-1` for that.

Reviewed By: sblackshear

Differential Revision: D3809394

fbshipit-source-id: cf2fdb184c593c7e7120c5d498994fb4f815d281