templates
Summary:
https://facebook.github.io/watchman/docs/bser.html

Deferring to the full doc^ which includes an example. Basically,
for an array of objects that have similar fields, record the field
names separately and then pack the objects.

The callback structure here is a little fine grained (separate
callbacks for fields inside of an object, inside of a template),
but it makes the json-building implementation relatively
straightforward.

Reviewed By: jopamer

Differential Revision: D3809424

fbshipit-source-id: bf4ac1ab97ca169c78230264c4f49a579865d465