Increase kMaxTrackedALocs to 256
Summary:
There's a number of units which require more than 128 tracked alias
locations, so increase the limit to 256. To reduce the impact of doubling the
bitset, introduce new utility functions to efficiently get the indices of only
the set bits. This reduces some of the overhead because we no longer have to
check each bit manually to see if its set or not.

Reviewed By: swtaarrs

Differential Revision: D4220777

fbshipit-source-id: 548c9f589c9d61358b75a5a087219e4af441694e