Extend unification to respect implied equality constraints
Summary: Unification doesn't currently respect "equality" constraints, either explicit (through `where`) or implicit (from `instanceof`). Equality constraints are represented as upper *and* lower bounds on type parameters in the type parameter environment, so unification needs to check against these if it tries to unify a type parameter T with a type t: in this case, it looks for t as an upper and lower bound of `T` (i.e. `T super t` and `T as t`).

Reviewed By: dlreeves, jamesjwu

Differential Revision: D3679402

fbshipit-source-id: 89f73283001abc22f670b3ec48094216d7fb64a5