Fix sign ext/mov vasms for PPC64.
Summary:
This commit fixes move and (signal) extensions vasms for PPC64. Some vasms
that needs to move values less than 64 bits to 64 bits registers. Those vasms were
translated to pairs that mov and sign extend (rlwinm + exts*) that are used even
when a sign extended are not intended like movz* or bitwise (xor*, and*)
which can lead to a wrong behaviour or emit a extra useless sign extend/mov instruction.

This commit also fixes some lowers that uses the same VReg in instructions
with operands that have different. Some lowered vasms are using movb and use
the result (VReg8) in operations that use VReg64 as input operands when we should
use movzbq in those cases.

To summarize this fix:

- Delete some redundant/wrong move instructions on sign extension instructions.
- Fix the problem with replacement by a nop when mov{r,r}. Some lowering instructions
like was using movb instructions, remove the instruction lead to invalid values
on registers. Replaced by movzbq.
- Replace movb/ext
Closes https://github.com/facebook/hhvm/pull/7455

Differential Revision: D4110281

Pulled By: mxw

fbshipit-source-id: db8c394b2d644879d41ca603f11a2bf340adbc35