Delete logging for lazy decl, clean up tests
Summary:
Since https://fburl.com/559135189 essentially everyone has been on lazy decl for a long time now. We can delete the logging for this code since it's on by default to all users now.

I also took this opportunity to clean up the LazyDecl tests, replacing them with the LazyInit tests instead and deleting them, which should decrease the total number of tests we have. I also pyfmt'ed the file so the lint errors are gone.

Reviewed By: dabek

Differential Revision: D4246703

fbshipit-source-id: 16d0486dc53213b957afd4bcf2976654681142f4