make enums a subtype of arraykey
Summary:
motivation:
1. keysets require their values to be arraykeys, and users often want to store enums in them.
2. ifirst-class enums in hack have a method called getNames() which returns an array keyed by the values. So, we require the values to be arraykeys already.

There is some argument to be made that opaque enums (i.e. without explicit constraint types) should not be arraykeys. However, that line of reasoning applies if we define arraykey as "int | string". if we instead define arraykey as "anything that can be used to index an array", then enums work as well.

thanks eshrews for the ocaml coaching, and andrewjkennedy dlreeves for suggesting different approaches to try out! this one ended up simplest and seems to work.

Reviewed By: andrewjkennedy

Differential Revision: D4258286

fbshipit-source-id: 45c8f64a02b1d510619167330cf2cef91896ad65