Make vasm-copy robust to statically-known physical reg displacements
Summary:
Before this diff, vasm-copy only tracked whether or not a physical
register was known to have the same value as it had on entry to the
program.  The goal of vasm-copy was to eliminate the virtual stack
pointer in non-resumed functions---but in the presence of inlining, when
the first inlined function call whose frame can't be eliminated will
modify rvmfp(), we would end up always materializing the frame pointer,
and requiring it for the rest of the Vunit.

We can do better by tracking the value of each reserved physical
register as a static displacement from the value of any other (or the
same) physical register on entry to the region.  This will let us
fold the def of a virtual register V into its use site, as long as at
both points, the physical register P relative to which V was defined is
a known displacement away from the entry-value of some other physical
register P'---since then we can just adjust by the difference in
displacements.

We also augment vasm-copy to track all reserved physical registers
(instead of just rvmfp() and rvmtl()), as well as to track the defs of
virtual registers in the analysis pass, so that if we def a virtual V
relative to a physical register P, then define another physical register
P' relative to V, we can propagate the tracked value of P to P'.

Reviewed By: markw65

Differential Revision: D4185784

fbshipit-source-id: a966524111eb6b6f29657dd270ef17ecaa93d273