Remove invert_grow_super and fix meth_caller
Summary:
During subtyping, we allow "unresolved" types to grow on the right of the subtype relation. That is, if we are solving
```
t <: a:=Tunresolved[t1,...,tn]
```
(where `a` is a type variable bound to a list of types in `Tunresolved`) then we let the list grow, so
```
a:=Tunresolved[t,t1,...,tn]
```

Now, if we are solving
```
C<t1> <: C<t2>
```
for a contravariant class `C`, then this reduces to
```
t2 <: t1
```
At present, we 'track" this flipping of the relation so that instead of growing unresolved types on
the right (in the supertype) we grow them on the left (in the subtype). This is broken (see additional test below).
There are other problems with subtyping (in particular, `a:=Tunresolved[] <: t` is ignored) and it's possible
that this diff will surface these, but let's fix the problems one at a time.

This fix also uncovered a problem with `meth_caller`. Now that function types are contravariant in parameters, we don't need to play the trick with `as` on an extra generic parameter that was used in the implementation of `meth_caller`. This fixes the problems with `meth_caller`: it is now sound.

Reviewed By: jamesjwu

Differential Revision: D4226678

fbshipit-source-id: 91e8f7fa8ec97cc04c08d1faec18bf3fe4dfb32c