Move Hack Full-fidelity-AST schema generation code to own directory.
Summary: I'm going to start moving more of the services associated with the full-fidelity parser into hh_client. One of those services will be "hey, what's the schema of the AST this version of hh_client is producing?"  To facilitate that, I'm moving the schema and associated transformation functions out of generate_full_fidelity.ml and into their own little library directory.

Reviewed By: dlreeves

Differential Revision: D4262405

fbshipit-source-id: c5fa5f20d1b05509de4554159affb26bf4d8b834