Support arbitrary IniSetting and ServerStats in access log
Summary:
This adds access log fields for inserting arbitrary IniSetting and ServerStats
values. I implemented it this way to make it easier to add other setting and
stat fields in the future. If there's a reason not to do it this way I can
implement the specific fields needed.

One issue I had is that user INI settings were cleared before access logging
was done, which meant this only reflected settings from the config file and not
changes made during requests. To fix this I moved
`IniSetting::ResetSavedDefaults()` from `requestShutdown()` of the extension
into `hphp_session_exit()`, but I'm not sure what the implications of this are.

This also adds a ServerStats key for `totalAlloc` named `mem.cumulative`.

Reviewed By: markw65

Differential Revision: D4273029

fbshipit-source-id: 89965bdbcffaf89cbca5a5a5979482a22b618004