Fix server command types
Summary:
This changes the inferred type of response from unit (which
was incorrect) to `_a'`, which is correct.

Pervasives "=" doesn't play well with GADTs in a
polymorphic context.

Pervasives "=" forces the LHS and RHS to be the same
since its signature is 'a -> 'a -> bool. This means
when you check "cmd = KILL", you force cmd to be
`unit ServerCommandTypes.t`, and response to be `unit`
which isn't actually correct.

No behavioral changes, just makes the output from Ocaml
merlin make sense. I'll be sending these down the debug
port so just want the types to be correct so it's easier for
me to do development

Reviewed By: dabek

Differential Revision: D4279670

fbshipit-source-id: 06a8086e6584d3d0f44f2c489d2a1f52c1acbb7a