Account for merged blocks when figuring out profile counts for region blocks
Summary:
Merged blocks can be created by guard relaxation.  For such blocks, we should
include the profile counts of all the merged blocks and not just one of the
blocks.  This diff adds a method to the RegionDesc to compute the proper
combined profile count, and uses it.  This required making the RegionDesc of the
current region being used to generate HHIR available deeper in irgen.  This is
done by adding a pointer to the current RegionDesc to IRGS.

Reviewed By: swtaarrs

Differential Revision: D4180710

fbshipit-source-id: 698f89d6a942e8accf82308ef738de728631a59e