Fix predicted inner type assertions while inlining
Summary:
When we're generating guards for a potential boxed type in a local,
stack slot, or the member base register, stop early if we know the guard will
always fail. If we don't, then we'll hit assertions while trying to access the
predicted inner type of the box (because its known to not be a box). Instead,
emit a Halt since this code should unreachable. This can happen during inlining
since we might have a more specific type for parameters.

Reviewed By: edwinsmith

Differential Revision: D4298341

fbshipit-source-id: 930e489464503b9010c78ca79e34f28af213f805