Change location of generated repo to outside the run directory.
Summary:
The run directory is erased on reboot.  This causes a problem with the oss-repo-mode script if the user has not already specified a repo location.  After a server reboot HHVM will fail to come up as it cannot find the repo.  Moving the default repo location into the cache folder should allow it to persist across reboots.

see https://github.com/facebook/hhvm/issues/7504

Reviewed By: aorenste

Differential Revision: D4322062

fbshipit-source-id: 20cae24a97567cf45b1bbf15a6282f61ab6dca0c