Allow deleting apache_note() values
Summary:
The default value for `$note_value` is currently an empty string, and this
determines the get only behaviour instead of set. This means that it is
impossible to clear a note in any way (can't be deleted, set to null or to an
empty string).

This changes `$note_value` to be nullable (but keeps the default as an empty
string), and deletes the note if `$note_value` is explicitly specified as null.
This is a bit weird, but when using notes in the access log deleting is better
than setting to an empty string.

Reviewed By: markw65

Differential Revision: D4301549

fbshipit-source-id: c15be06b77cb8910e853ed777ba1a8e17fa53fe2