HfsortPlus layout algorithm
Summary:
This is a new hfsort+ algorithm that layouts hot functions respecting iTLB cache. I may adjust some constants but the general implementation is ready.

The general idea is to optimize the function given by expectedCacheHitRatio in hfsort_plus.cpp, which is an expected cache hit ratio under the assumption that the functions are sent to iTLB cache in a random order and respect the observed frequencies (i.e., samples and edge weights). See also comments in the code.

The algorithm takes under a minute for a single run, which is faster than reading perf data but slower than the default hfsort. Though we may speed it up, if needed.

Reviewed By: ottoni

Differential Revision: D4230591

fbshipit-source-id: 9413565812bdeab0434d4d18078d6fea0fe76758