Make attribute value list parsing use helper methods
Summary:
Attribute value list parsing is now consistent with other list parsing.

This is part of a series of diffs aimed at first getting all the list processing in the FFP consistently using helper methods, and second, modifying the behaviour of those helpers to produce the best list representation for automated tools.

Reviewed By: jopamer

Differential Revision: D4313570

fbshipit-source-id: 4165fc0b204c4d72b1aec851cbe7da80a3c15323