Log perf warnings when checkTmpsSpanningCalls() fails instead of crashing
Summary:
We've been capable of correctly handling values live across calls for
years now, but we kept this assertion around because it usually indicated that
something worth looking into was going on. This situation is exactly what perf
warnings are for, so do that instead of crashing.

I renamed the "ir_unit" column in the alias-analysis perf warning to
"hhir_unit", to be consistent with the new one. I think it's important to be
specific here, especially since there's also a vasm_unit column for xls punts.

Reviewed By: ricklavoie

Differential Revision: D4218244

fbshipit-source-id: 2c481f0aaf45436c8f70ba4a18b1497384f3a7f1