Clean up list processing for "for" statement parsing
Summary:
Continuing with my stack of diffs making use of helper methods to make parsing lists shorter and more consistent.

In this diff, the for-loop now uses helper methods. Note that previously for something like

    for ($x = 1, $y = 1 ; ...

we constructed the list EXPRESSION, COMMA, EXPRESSION.  This is inconsistent with practice throughout the rest of the parser. We now construct  LISTITEM(EXPRESSION, COMMA), EXPRESSION.

Eventually we'll construct the list LISTITEM(EXPRESSION, COMMA), LISTITEM(EXPRESSION, MISSING).  But that will come in a later diff.

Reviewed By: jopamer

Differential Revision: D4323235

fbshipit-source-id: e8637d8869384367ba0168579da0f0bc4a1c4a52