Typed AST: conditional and null-coalescing expressions
Summary:
Annotate `Eif` expressions. Null coalescing expressions are a derived form (they are already type-checked using the same code).

In future, we should make type refinement explicit, so simplify the type-checking of the typed AST and downstream IR. For example, to check
```
$x instanceof C ? $x->foo() : $y
```
we need to know about refinement of `$x` in the `$x->foo()` expression - the types of the locals have been "updated".

Reviewed By: dlreeves

Differential Revision: D4357487

fbshipit-source-id: 74d3e66fd89d4aac66bb9eb2b3d425f4a5fd9a80