Iterate in address order, avoid sorting PtrMap
Summary:
Before a collection starts, we initialize a PtrMap with all
the ranges of heap objects, then sort it, which can be expensive
since PtrMap can be large (>1M entries). However, most of the
entries in PtrMap come from slabs, so if we sort slabs (and big
objects) first, then iterate, we can build PtrMap in address order.

Reviewed By: kodafb

Differential Revision: D4348514

fbshipit-source-id: e2e21321e1e1508a72af26df6526ebac536a272f