bad assumption about proc_nice results
Summary:
proc_nice_basic.php is the only test which depends on the relative priority difference
before and after a call to proc_nice().  The other tests in the suite make sure the
function call accepts various parameters.

The test was passing on the CI @ http://hhvm-arm64.theobroma-systems.com/index-fb.html,
4 core machine, but consistently failing in our internal runs on 48-core machines.
It would pass when run by hand, so I suspected it had something to do with system load.

After instrumenting a version of the test to print out more information it became clear
that the test was assuming that a call to proc_nice would always change priority by a known
delta.  This is not the case if this test or another one that calls proc_nice has already been
run.  It just happened in our environment that one of the hhvm servers already run a test
with a call to proc_nice, so the starting priority was lower.

I was able to demonstrate this failure with the following command which runs the test
on a single thread mu
Closes https://github.com/facebook/hhvm/pull/7569

Differential Revision: D4371073

Pulled By: mofarrell

fbshipit-source-id: dceaa39b1bd057fb403113e1861e2818d128e196