Parse lists consistently with expectations, rather than grammar
Summary:
There are two major changes here to how lists are represented in the FF parser. Both will make it more consistent to process lists in analysis / refactoring tools.

When I designed the list processing code for the parser I matched (some of!) the conventions of a grammar. For example, we might see

    compound-statement :
      { statement-list-opt }
    statement-list :
      statement
      statement-list statement

The intent of this is clear; the body of the compound statement can be one of three things, grammatically: empty, a single statement, or a sequence of two or more statements.

The way I represented this was: if the body is empty then we have a MISSING node. If it is a single statement then we have a STATEMENT node. If it is two or more statements then we have a LIST ( STATEMENT; STATEMENT; STATEMENT ... ) node.

Though this is conceptually quite nice, it's a little bit weird for tools that must process this format to have to consider these three cases.

So the first big change is: an empty list is still MISSING.  A non-empty list is a LIST ( STATEMENT; ... ) even for singletons.

We might consider going all the way later and saying that the empty list is LIST() with no children. I'll consider that for a future diff.

The second big change involves "separated" lists. Consider a comma-separated list of foos which allows a trailing comma:

     foo-list:
        (   foos  ,-opt  )
     foos :
       foo
       foos  ,  foo

We used to represent the list inside the parens as follows:

    ( )           --   MISSING
    ( foo )       --   FOO
    ( foo, )      --   LIST_ITEM(FOO, COMMA)
    ( foo, foo )  --   LIST(LIST_ITEM(FOO, COMMA), FOO)
    ( foo, foo, ) --   LIST(LIST_ITEM(FOO, COMMA), LIST_ITEM(FOO, COMMA))

So first, again we have this business of its only a list when there are two things in it. And second, it's only a list item when there's a comma.  The change is to say: in a list where separators are expected, every element will be a list item consistently:

    ( )           --   MISSING
    ( foo )       --   LIST(LIST_ITEM(FOO, MISSING))
    ( foo, )      --   LIST(LIST_ITEM(FOO, COMMA))
    ( foo, foo )  --   LIST(LIST_ITEM(FOO, COMMA), LIST_ITEM(FOO, MISSING))
    ( foo, foo, ) --   LIST(LIST_ITEM(FOO, COMMA), LIST_ITEM(FOO, COMMA))

Note that the last item will be a list-item with missing comma even for lists that are not allowed to have a trailing comma.  Missing does not mean "error", it just means "not in the source".

Differential Revision: D4341216

fbshipit-source-id: 7b69ac007af9a8b33a37b99a68023d6fb3b0f008